Remove cgo from coverage test (#1483)
I no longer believe there is a way to consistently collect coverage
data for C libraries with Bazel. This test kept getting stuck in good
or bad states, and its success depended on other tests that were
running with it and tools (lcov) installed on the system.

cgo coverage won't be supported until this is fixed in Bazel.

Also:

Deleted coverage_bin_test since it's expensive to run and provides
very little value.