Add more explicit dependencies for bazel_test (#1724)
//tests:all_rules_md5 now includes .bzl files in //proto, //proto/wkt,
and .go files in //go/tools/builders.

all_rules_md5 is an implicit data dependency of bazel_test
rules. These tests verify the rules_go implementation in an alternate
workspace, so their dependencies on .bzl files in the current
workspace is implicit. When all_rules_md5 changes, it forces them to
run again.